web/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "dynaconf>=3.2.10",
    "ffmpeg-python>=0.2.0",
    "flask>=3.1.0",
    "flask-jwt-extended>=4.7.1",
    "flask-sqlalchemy>=3.1.1",
    "joblib>=1.4.2",
    "numpy>=2.2.4",
    "openai>=1.72.0",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pymysql>=1.1.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.40",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "waitress>=3.0.2",
    "aisearch @ file://${PROJECT_ROOT}/AISearch",
    "langchain-huggingface>=0.2.0",
    "playwright>=1.52.0",
    "redis",
    "pytest>=8.4.1",
]
[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
